Added save to csv, the ability to communicate with host, little work done on filtering, legend for chart now shows below the chart.

+++ tool/cpp/event.h
@@ -29,7 +29,7 @@ namespace Event
 
     struct Event {
         Event();
-        Event(uint64_t eventNumber, unsigned int eventType, uint64_t timestamp)
+        Event(uint64_t eventNumber, size_t eventType, uint64_t timestamp)
         : eventNumber(eventNumber), eventType(eventType), timestamp(timestamp) {};
         ~Event() {};
 
@@ -40,12 +40,12 @@ namespace Event
 	    virtual void getAsVerbose(std::stringstream& ss) = 0;
 
         uint64_t eventNumber;
-        unsigned int eventType;
+        size_t eventType;
         uint64_t timestamp;
     };
 
     struct InitStream : public Event {
-        InitStream(uint64_t eventNumber, unsigned int eventType, uint64_t timestamp, uint64_t stream_magic, std::string& platform, uint64_t frequency)
+        InitStream(uint64_t eventNumber, size_t eventType, uint64_t timestamp, uint64_t stream_magic, std::string& platform, uint64_t frequency)
         : Event(eventNumber, eventType, timestamp), stream_magic(stream_magic), platform(platform), system_frequency(system_frequency)
         { };
         ~InitStream();
@@ -69,7 +69,7 @@ namespace Event
     };
 
     struct StopStream : public Event {
-        StopStream(uint64_t eventNumber, unsigned int eventType, uint64_t timestamp)
+        StopStream(uint64_t eventNumber, size_t eventType, uint64_t timestamp)
         : Event(eventNumber, eventType, timestamp)
         { };
         ~StopStream();
@@ -84,7 +84,7 @@ namespace Event
     };
 
     struct AddHeap : public Event {
-        AddHeap(uint64_t eventNumber, unsigned int eventType, uint64_t time, uint64_t id, std::string& type, std::string& name)
+        AddHeap(uint64_t eventNumber, size_t eventType, uint64_t time, uint64_t id, std::string& type, std::string& name)
         : Event(eventNumber, eventType, time)
         , id(id)
         , type(type)
@@ -111,7 +111,7 @@ namespace Event
     };
 
     struct RemoveHeap : public Event {
-        RemoveHeap(uint64_t eventNumber, unsigned int eventType, uint64_t time, uint64_t id)
+        RemoveHeap(uint64_t eventNumber, size_t eventType, uint64_t time, uint64_t id)
         : Event(eventNumber, eventType, time), id(id) {};
         ~RemoveHeap();
 
@@ -129,7 +129,7 @@ namespace Event
     };
 
     struct AddCore : public Event {
-        AddCore(uint64_t eventNumber, unsigned int eventType, uint64_t time, uint64_t id,uint64_t pointer,uint64_t size_bytes)
+        AddCore(uint64_t eventNumber, size_t eventType, uint64_t time, uint64_t id,uint64_t pointer,uint64_t size_bytes)
         : Event(eventNumber, eventType, time), id(id), pointer(pointer), size(size_bytes) { };
         ~AddCore();
 
@@ -152,7 +152,7 @@ namespace Event
     };
 
     struct RemoveCore : public Event {
-        RemoveCore(uint64_t eventNumber, unsigned int eventType, uint64_t time, uint64_t id,uint64_t pointer,uint64_t size_bytes)
+        RemoveCore(uint64_t eventNumber, size_t eventType, uint64_t time, uint64_t id,uint64_t pointer,uint64_t size_bytes)
         : Event(eventNumber, eventType, time), id(id), pointer(pointer), size(size_bytes) { };
         ~RemoveCore();
 
@@ -175,7 +175,7 @@ namespace Event
     };
 
     struct AddAllocation : public Event {
-        AddAllocation(uint64_t eventNumber, unsigned int eventType, uint64_t time, uint64_t id,uint64_t pointer,uint64_t size_bytes)
+        AddAllocation(uint64_t eventNumber, size_t eventType, uint64_t time, uint64_t id,uint64_t pointer,uint64_t size_bytes)
         : Event(eventNumber, eventType, time), id(id), pointer(pointer), size(size_bytes) { };
         ~AddAllocation();
 
@@ -198,7 +198,7 @@ namespace Event
     };
 
     struct RemoveAllocation : public Event {
-        RemoveAllocation(uint64_t eventNumber, unsigned int eventType, uint64_t time, uint64_t id,uint64_t pointer)
+        RemoveAllocation(uint64_t eventNumber, size_t eventType, uint64_t time, uint64_t id,uint64_t pointer)
         : Event(eventNumber, eventType, time), id(id), pointer(pointer) { };
         ~RemoveAllocation();
 
@@ -219,7 +219,7 @@ namespace Event
     };
 
     struct StartEvent : public Event {
-        StartEvent(uint64_t eventNumber, unsigned int eventType, uint64_t time, std::string& name)
+        StartEvent(uint64_t eventNumber, size_t eventType, uint64_t time, std::string& name)
         : Event(eventNumber, eventType, time), name(name) { };
         ~StartEvent();
 
@@ -237,7 +237,7 @@ namespace Event
     };
 
     struct EndEvent : public Event {
-        EndEvent(uint64_t eventNumber, unsigned int eventType, uint64_t time, std::string& name)
+        EndEvent(uint64_t eventNumber, size_t eventType, uint64_t time, std::string& name)
         : Event(eventNumber, eventType, time), name(name) { };
         ~EndEvent();
 
@@ -255,7 +255,7 @@ namespace Event
     };
 
     struct HeapSetBackingAllocator : public Event {
-        HeapSetBackingAllocator(uint64_t eventNumber, unsigned int eventType, uint64_t time, size_t for_heap, size_t set_to_heap)
+        HeapSetBackingAllocator(uint64_t eventNumber, size_t eventType, uint64_t time, size_t for_heap, size_t set_to_heap)
         : Event(eventNumber, eventType, time), for_heap(for_heap), set_to_heap(set_to_heap) { };
         ~HeapSetBackingAllocator();
 
@@ -270,8 +270,8 @@ namespace Event
             ss << "\n\tset to heap: " << set_to_heap << "\n";
         }
 
-        unsigned int for_heap;
-        unsigned int set_to_heap;
+        size_t for_heap;
+        size_t set_to_heap;
     };
 }
 #endif //EVENT_H
